+++ ext/db/captcha.go
@@ -4,23 +4,17 @@ import (
 	"database/sql"
 )
 
-const insertIsbn = `
-INSERT INTO captcha (isbn)
-VALUES ($1)
-ON CONFLICT DO NOTHING`
+const insertIsbn = `INSERT INTO captcha (isbn) VALUES ($1) ON CONFLICT DO NOTHING`
 
-const removeIsbn = `
-DELETE FROM captcha
-WHERE isbn = $1`
+const removeIsbn = `DELETE FROM captcha WHERE isbn = $1`
 
-const listIsbn = `
-SELECT isbn FROM captcha`
+const listIsbn = `SELECT isbn FROM captcha`
 
-func (db *DB) CaptchaAdd(isbn string) (int, error) {
+func (db *DB) CaptchaAdd(isbn string) (int64, error) {
 	return db.Exec(insertIsbn, isbn)
 }
 
-func (db *DB) CaptchaRemove(isbn string) (int, error) {
+func (db *DB) CaptchaRemove(isbn string) (int64, error) {
 	return db.Exec(removeIsbn, isbn)
 }
